ref(rust): Set up logging integration correctly
Switch to pretty-env-logger, since that's what we have docs for: https://docs.rs/sentry/0.31.7/sentry/integrations/log/index.html
